Moved routes to a separate file where they can be tested.
